Refactor: add page title in useTitle hook & move DecisionsPreview component
