+++ src/modules/dashboard/components/DecisionPreview/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DecisionPreview';
