+++ src/utils/hooks/useTitle.ts
@@ -24,6 +24,7 @@ import {
   USER_ROUTE,
   MEMBERS_ROUTE,
   COLONY_DECISIONS_ROUTE,
+  COLONY_DECISIONS_PREVIEW_ROUTE,
   COLONY_EXTENSION_SETUP_ROUTE,
   COLONY_EXTENSION_DETAILS_ROUTE,
   COLONY_EXTENSIONS_ROUTE,
@@ -92,6 +93,11 @@ const MSG = defineMessages({
     defaultMessage: `Decisions | Colony - {colonyName}`,
   },
 
+  decisionPreview: {
+    id: 'utils.hooks.useTitle.decisionPreview',
+    defaultMessage: `Decision Preview | Colony - {colonyName}`,
+  },
+
   colonyExtensions: {
     id: 'utils.hooks.useTitle.colonyExtensions',
     defaultMessage: `Extensions | Colony - {colonyName}`,
@@ -157,6 +163,7 @@ const routeMessages: Record<string, MessageDescriptor> = {
   [COLONY_EVENTS_ROUTE]: MSG.colonyEvents,
   [COLONY_FUNDING_ROUTE]: MSG.colonyFunds,
   [COLONY_DECISIONS_ROUTE]: MSG.colonyDecisions,
+  [COLONY_DECISIONS_PREVIEW_ROUTE]: MSG.decisionPreview,
   [COLONY_EXTENSIONS_ROUTE]: MSG.colonyExtensions,
   [COLONY_EXTENSION_DETAILS_ROUTE]: MSG.colonyExtensionDetails,
   [COLONY_EXTENSION_SETUP_ROUTE]: MSG.colonyExtensionSetup,
